move platform discard/commit from ManifestManager to DesignManager
