+++ src/main/java/com/oneops/transistor/service/DesignManager.java
@@ -24,11 +24,16 @@ import com.oneops.cms.dj.domain.CmsRfcCI;
 import com.oneops.transistor.domain.CatalogExport;
 import com.oneops.transistor.export.domain.DesignExportSimple;
 
+import java.util.List;
+
 @Transactional
 public interface DesignManager {
 	public long generatePlatform(CmsRfcCI platRfc, long assemblyId, String userId, String scope);
 	public long deletePlatform(long platformId, String userId, String scope);
 	public long clonePlatform(CmsRfcCI platRfc, Long targetAssemblyId, long sourcePlatId, String userId, String scope);
+	public List<CmsRfcCI> getPlatformRfcs(long platId, String userId);
+	public long discardReleaseForPlatform(long platId, String userId);
+	public long commitReleaseForPlatform(long platId, String desc, String userId);
 	public long cloneAssembly(CmsCI assemblyCI, long sourceAssemblyId, String userId, String scope);
 	public long saveAssemblyAsCatalog(CmsCI catalogCI, long sourceAssemblyId, String userId, String scope);
 	public CatalogExport exportCatalog(long catalogCIid, String scope);
