+++ src/main/java/com/oneops/transistor/service/DesignRfcProcessor.java
@@ -23,6 +23,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import com.oneops.cms.dj.service.CmsRfcProcessor;
 import org.apache.log4j.Logger;
 
 import com.oneops.cms.cm.domain.CmsCI;
@@ -49,9 +50,14 @@ public class DesignRfcProcessor {
 	private TransUtil trUtil;
 	private CmsCmProcessor cmProcessor;
 	private CmsCmRfcMrgProcessor cmRfcMrgProcessor;
+    private CmsRfcProcessor rfcProcessor;
 
 
-	public void setTrUtil(TransUtil trUtil) {
+    public void setRfcProcessor(CmsRfcProcessor rfcProcessor) {
+        this.rfcProcessor = rfcProcessor;
+    }
+
+    public void setTrUtil(TransUtil trUtil) {
 		this.trUtil = trUtil;
 	}
 	
@@ -544,4 +550,29 @@ public class DesignRfcProcessor {
 		return newRfc;
 	}
 
+
+    private static String getPlatformNs(CmsRfcCI platform) {
+        return platform.getNsPath()+"/_design/"+platform.getCiName();
+    }
+
+
+    public List<CmsRfcCI> getPlatformRfcs(long platId) {
+        String nsPath = getPlatformNs(cmRfcMrgProcessor.getCiById(platId));
+        return rfcProcessor.getOpenRfcCIByClazzAndNameNoAttrs(nsPath, null, null);
+    }
+
+    public long discardReleaseForPlatform(long platId) {
+        CmsRfcCI platformRfc= cmRfcMrgProcessor.getCiById(platId);
+        rfcProcessor.rmRfcs(getPlatformNs(platformRfc));
+        if (platformRfc.getIsActiveInRelease()){
+            rfcProcessor.rmRfcCiFromRelease(platformRfc.getRfcId());
+        }
+        return platId;
+    }
+
+
+    public long commitReleaseForPlatform(long platId, String desc, String userId) {
+        CmsRfcCI platformRfc= cmRfcMrgProcessor.getCiById(platId);
+        return rfcProcessor.commitReleaseForPlatform(platformRfc, desc, userId);
+    }
 }
