+++ src/main/webapp/WEB-INF/application-context.xml
@@ -42,6 +42,7 @@
 		<property name="cmRfcMrgProcessor" ref="cmsCmRfcMrgProcessor" />
 		<property name="cmProcessor" ref="cmsCmProcessor" />
 		<property name="mdProcessor" ref="cmsMdProcessor" />
+		<property name="rfcProcessor" ref="cmsRfcProcessor" />
         <property name="trUtil" ref="trUtil" />
     </bean>
